requirements.txt
python-telegram-bot==22.5
anthropic>=0.69.0
asyncpg==0.31.0
